+++ universal-application-tool-0.0.1/app/views/AwsFileUploadViewStrategy.java
@@ -5,6 +5,9 @@ import static j2html.TagCreator.each;
 import static j2html.TagCreator.form;
 import static j2html.TagCreator.input;
 import static j2html.attributes.Attr.ENCTYPE;
+import static j2html.attributes.Attr.FORM;
+import static views.BaseHtmlView.makeCsrfTokenInputTag;
+import static views.BaseHtmlView.submitButton;
 
 import controllers.applicant.routes;
 import j2html.attributes.Attr;
@@ -13,6 +16,7 @@ import j2html.tags.Tag;
 import java.util.Optional;
 import play.i18n.Messages;
 import play.mvc.Http.HttpVerbs;
+import services.MessageKey;
 import services.applicant.question.FileUploadQuestion;
 import services.cloud.FileNameFormatter;
 import services.cloud.StorageUploadRequest;
@@ -20,6 +24,7 @@ import services.cloud.aws.SignedS3UploadRequest;
 import views.applicant.ApplicantProgramBlockEditView.Params;
 import views.questiontypes.ApplicantQuestionRendererFactory;
 import views.questiontypes.ApplicantQuestionRendererParams;
+import views.style.ApplicantStyles;
 import views.style.BaseStyles;
 import views.style.ReferenceClasses;
 import views.style.Styles;
@@ -106,6 +111,110 @@ public class AwsFileUploadViewStrategy extends FileUploadViewStrategy {
     Tag skipForms = renderDeleteAndContinueFileUploadForms(params);
     Tag buttons = renderFileUploadBottomNavButtons(params);
     return div(uploadForm, skipForms, buttons);
+  }
+
+  Tag renderFileUploadBottomNavButtons(Params params) {
+    Optional<Tag> maybeContinueButton = maybeRenderContinueButton(params);
+    Optional<Tag> maybeSkipOrDeleteButton = maybeRenderSkipOrDeleteButton(params);
+    ContainerTag ret =
+        div()
+            .withClasses(ApplicantStyles.APPLICATION_NAV_BAR)
+            // An empty div to take up the space to the left of the buttons.
+            .with(div().withClasses(Styles.FLEX_GROW))
+            .with(renderReviewButton(params));
+    if (maybeSkipOrDeleteButton.isPresent()) {
+      ret.with(maybeSkipOrDeleteButton.get());
+    }
+    ret.with(renderUploadButton(params));
+    if (maybeContinueButton.isPresent()) {
+      ret.with(maybeContinueButton.get());
+    }
+    return ret;
+  }
+
+  /**
+   * Renders a form submit button for continue form if an uploaded file is present.
+   *
+   * <p>See {@link renderDeleteAndContinueFileUploadForms}.
+   */
+  Optional<Tag> maybeRenderContinueButton(Params params) {
+    if (!hasUploadedFile(params)) {
+      return Optional.empty();
+    }
+    Tag button =
+        submitButton(params.messages().at(MessageKey.BUTTON_KEEP_FILE.getKeyName()))
+            .attr(FORM, FILEUPLOAD_CONTINUE_FORM_ID)
+            .withClasses(ApplicantStyles.BUTTON_BLOCK_NEXT)
+            .withId(FILEUPLOAD_CONTINUE_BUTTON_ID);
+    return Optional.of(button);
+  }
+
+  /**
+   * Renders a form submit button for delete form if the file upload question is optional.
+   *
+   * <p>If an uploaded file is present, render the button text as delete. Otherwise, skip.
+   *
+   * <p>See {@link renderDeleteAndContinueFileUploadForms}.
+   */
+  private Optional<Tag> maybeRenderSkipOrDeleteButton(Params params) {
+    if (hasAtLeastOneRequiredQuestion(params)) {
+      // If the file question is required, skip or delete is not allowed.
+      return Optional.empty();
+    }
+    String buttonText = params.messages().at(MessageKey.BUTTON_SKIP_FILEUPLOAD.getKeyName());
+    String buttonId = FILEUPLOAD_SKIP_BUTTON_ID;
+    if (hasUploadedFile(params)) {
+      buttonText = params.messages().at(MessageKey.BUTTON_DELETE_FILE.getKeyName());
+      buttonId = FILEUPLOAD_DELETE_BUTTON_ID;
+    }
+    Tag button =
+        submitButton(buttonText)
+            .attr(FORM, FILEUPLOAD_DELETE_FORM_ID)
+            .withClasses(ApplicantStyles.BUTTON_REVIEW)
+            .withId(buttonId);
+    return Optional.of(button);
+  }
+
+  /**
+   * Returns two hidden forms for navigating through a file upload block without uploading a file.
+   *
+   * <p>Delete form sends an update with an empty file key. An empty file key erases the existing
+   * file key if one is present. In either case, the file upload question is marked as seen but
+   * unanswered, namely skipping the file upload. This is only allowed for an optional question.
+   *
+   * <p>Continue form sends an update with the currently stored file key, the same behavior as an
+   * applicant re-submits a form without changing their answer. Continue form is only used when an
+   * existing file (and file key) is present.
+   */
+  Tag renderDeleteAndContinueFileUploadForms(Params params) {
+    String formAction =
+        routes.ApplicantProgramBlocksController.update(
+                params.applicantId(), params.programId(), params.block().getId(), params.inReview())
+            .url();
+    ApplicantQuestionRendererParams rendererParams =
+        ApplicantQuestionRendererParams.builder().setMessages(params.messages()).build();
+
+    Tag continueForm =
+        form()
+            .withId(FILEUPLOAD_CONTINUE_FORM_ID)
+            .withAction(formAction)
+            .withMethod(HttpVerbs.POST)
+            .with(makeCsrfTokenInputTag(params.request()))
+            .with(
+                each(
+                    params.block().getQuestions(),
+                    question -> renderFileKeyField(question, rendererParams)));
+    Tag deleteForm =
+        form()
+            .withId(FILEUPLOAD_DELETE_FORM_ID)
+            .withAction(formAction)
+            .withMethod(HttpVerbs.POST)
+            .with(makeCsrfTokenInputTag(params.request()))
+            .with(
+                each(
+                    params.block().getQuestions(),
+                    question -> renderEmptyFileKeyField(question, rendererParams)));
+    return div(continueForm, deleteForm).withClasses(Styles.HIDDEN);
   }
 
   private String acceptFileTypes() {
