+++ universal-application-tool-0.0.1/app/views/FileUploadViewStrategy.java
@@ -1,19 +1,13 @@
 package views;
 
 import static j2html.TagCreator.a;
-import static j2html.TagCreator.div;
-import static j2html.TagCreator.each;
-import static j2html.TagCreator.form;
 import static j2html.attributes.Attr.FORM;
 import static j2html.attributes.Attr.HREF;
-import static views.BaseHtmlView.makeCsrfTokenInputTag;
 import static views.BaseHtmlView.submitButton;
 
 import controllers.applicant.routes;
 import j2html.tags.ContainerTag;
 import j2html.tags.Tag;
-import java.util.Optional;
-import play.mvc.Http.HttpVerbs;
 import services.MessageKey;
 import services.applicant.question.ApplicantQuestion;
 import services.applicant.question.FileUploadQuestion;
@@ -22,7 +16,6 @@ import views.questiontypes.ApplicantQuestionRendererFactory;
 import views.questiontypes.ApplicantQuestionRendererParams;
 import views.questiontypes.FileUploadQuestionRenderer;
 import views.style.ApplicantStyles;
-import views.style.Styles;
 
 /**
  * A strategy pattern that abstracts out the logic of file upload/download into the different cloud
@@ -32,13 +25,13 @@ public abstract class FileUploadViewStrategy {
 
   static final String MIME_TYPES_IMAGES_AND_PDF = "image/*,.pdf";
   final String BLOCK_FORM_ID = "cf-block-form";
-  private final String FILEUPLOAD_CONTINUE_FORM_ID = "cf-fileupload-continue-form";
-  private final String FILEUPLOAD_DELETE_FORM_ID = "cf-fileupload-delete-form";
-  private final String FILEUPLOAD_SUBMIT_FORM_ID = "cf-block-submit";
-  private final String FILEUPLOAD_DELETE_BUTTON_ID = "fileupload-delete-button";
-  private final String FILEUPLOAD_SKIP_BUTTON_ID = "fileupload-skip-button";
-  private final String FILEUPLOAD_CONTINUE_BUTTON_ID = "fileupload-continue-button";
-  private final String REVIEW_APPLICATION_BUTTON_ID = "review-application-button";
+  final String FILEUPLOAD_CONTINUE_FORM_ID = "cf-fileupload-continue-form";
+  final String FILEUPLOAD_DELETE_FORM_ID = "cf-fileupload-delete-form";
+  final String FILEUPLOAD_SUBMIT_FORM_ID = "cf-block-submit";
+  final String FILEUPLOAD_DELETE_BUTTON_ID = "fileupload-delete-button";
+  final String FILEUPLOAD_SKIP_BUTTON_ID = "fileupload-skip-button";
+  final String FILEUPLOAD_CONTINUE_BUTTON_ID = "fileupload-continue-button";
+  final String REVIEW_APPLICATION_BUTTON_ID = "review-application-button";
 
   /**
    * Method to generate the field tags for the file upload view form.
@@ -60,48 +53,6 @@ public abstract class FileUploadViewStrategy {
   public abstract Tag renderFileUploadBlockSubmitForms(
       Params params, ApplicantQuestionRendererFactory applicantQuestionRendererFactory);
 
-  /**
-   * Returns two hidden forms for navigating through a file upload block without uploading a file.
-   *
-   * <p>Delete form sends an update with an empty file key. An empty file key erases the existing
-   * file key if one is present. In either case, the file upload question is marked as seen but
-   * unanswered, namely skipping the file upload. This is only allowed for an optional question.
-   *
-   * <p>Continue form sends an update with the currently stored file key, the same behavior as an
-   * applicant re-submits a form without changing their answer. Continue form is only used when an
-   * existing file (and file key) is present.
-   */
-  Tag renderDeleteAndContinueFileUploadForms(Params params) {
-    String formAction =
-        routes.ApplicantProgramBlocksController.update(
-                params.applicantId(), params.programId(), params.block().getId(), params.inReview())
-            .url();
-    ApplicantQuestionRendererParams rendererParams =
-        ApplicantQuestionRendererParams.builder().setMessages(params.messages()).build();
-
-    Tag continueForm =
-        form()
-            .withId(FILEUPLOAD_CONTINUE_FORM_ID)
-            .withAction(formAction)
-            .withMethod(HttpVerbs.POST)
-            .with(makeCsrfTokenInputTag(params.request()))
-            .with(
-                each(
-                    params.block().getQuestions(),
-                    question -> renderFileKeyField(question, rendererParams)));
-    Tag deleteForm =
-        form()
-            .withId(FILEUPLOAD_DELETE_FORM_ID)
-            .withAction(formAction)
-            .withMethod(HttpVerbs.POST)
-            .with(makeCsrfTokenInputTag(params.request()))
-            .with(
-                each(
-                    params.block().getQuestions(),
-                    question -> renderEmptyFileKeyField(question, rendererParams)));
-    return div(continueForm, deleteForm).withClasses(Styles.HIDDEN);
-  }
-
   Tag renderQuestion(
       ApplicantQuestion question,
       ApplicantQuestionRendererParams params,
@@ -109,36 +60,15 @@ public abstract class FileUploadViewStrategy {
     return applicantQuestionRendererFactory.getRenderer(question).render(params);
   }
 
-  private Tag renderFileKeyField(
-      ApplicantQuestion question, ApplicantQuestionRendererParams params) {
+  Tag renderFileKeyField(ApplicantQuestion question, ApplicantQuestionRendererParams params) {
     return FileUploadQuestionRenderer.renderFileKeyField(question, params, false);
   }
 
-  private Tag renderEmptyFileKeyField(
-      ApplicantQuestion question, ApplicantQuestionRendererParams params) {
+  Tag renderEmptyFileKeyField(ApplicantQuestion question, ApplicantQuestionRendererParams params) {
     return FileUploadQuestionRenderer.renderFileKeyField(question, params, true);
   }
 
-  Tag renderFileUploadBottomNavButtons(Params params) {
-    Optional<Tag> maybeContinueButton = maybeRenderContinueButton(params);
-    Optional<Tag> maybeSkipOrDeleteButton = maybeRenderSkipOrDeleteButton(params);
-    ContainerTag ret =
-        div()
-            .withClasses(ApplicantStyles.APPLICATION_NAV_BAR)
-            // An empty div to take up the space to the left of the buttons.
-            .with(div().withClasses(Styles.FLEX_GROW))
-            .with(renderReviewButton(params));
-    if (maybeSkipOrDeleteButton.isPresent()) {
-      ret.with(maybeSkipOrDeleteButton.get());
-    }
-    ret.with(renderUploadButton(params));
-    if (maybeContinueButton.isPresent()) {
-      ret.with(maybeContinueButton.get());
-    }
-    return ret;
-  }
-
-  private Tag renderReviewButton(Params params) {
+  Tag renderReviewButton(Params params) {
     String reviewUrl =
         routes.ApplicantProgramReviewController.review(params.applicantId(), params.programId())
             .url();
@@ -148,50 +78,7 @@ public abstract class FileUploadViewStrategy {
         .withClasses(ApplicantStyles.BUTTON_REVIEW);
   }
 
-  /**
-   * Renders a form submit button for continue form if an uploaded file is present.
-   *
-   * <p>See {@link renderDeleteAndContinueFileUploadForms}.
-   */
-  private Optional<Tag> maybeRenderContinueButton(Params params) {
-    if (!hasUploadedFile(params)) {
-      return Optional.empty();
-    }
-    Tag button =
-        submitButton(params.messages().at(MessageKey.BUTTON_KEEP_FILE.getKeyName()))
-            .attr(FORM, FILEUPLOAD_CONTINUE_FORM_ID)
-            .withClasses(ApplicantStyles.BUTTON_BLOCK_NEXT)
-            .withId(FILEUPLOAD_CONTINUE_BUTTON_ID);
-    return Optional.of(button);
-  }
-
-  /**
-   * Renders a form submit button for delete form if the file upload question is optional.
-   *
-   * <p>If an uploaded file is present, render the button text as delete. Otherwise, skip.
-   *
-   * <p>See {@link renderDeleteAndContinueFileUploadForms}.
-   */
-  private Optional<Tag> maybeRenderSkipOrDeleteButton(Params params) {
-    if (hasAtLeastOneRequiredQuestion(params)) {
-      // If the file question is required, skip or delete is not allowed.
-      return Optional.empty();
-    }
-    String buttonText = params.messages().at(MessageKey.BUTTON_SKIP_FILEUPLOAD.getKeyName());
-    String buttonId = FILEUPLOAD_SKIP_BUTTON_ID;
-    if (hasUploadedFile(params)) {
-      buttonText = params.messages().at(MessageKey.BUTTON_DELETE_FILE.getKeyName());
-      buttonId = FILEUPLOAD_DELETE_BUTTON_ID;
-    }
-    Tag button =
-        submitButton(buttonText)
-            .attr(FORM, FILEUPLOAD_DELETE_FORM_ID)
-            .withClasses(ApplicantStyles.BUTTON_REVIEW)
-            .withId(buttonId);
-    return Optional.of(button);
-  }
-
-  private Tag renderUploadButton(Params params) {
+  Tag renderUploadButton(Params params) {
     String styles = ApplicantStyles.BUTTON_BLOCK_NEXT;
     if (hasUploadedFile(params)) {
       styles = ApplicantStyles.BUTTON_REVIEW;
@@ -202,14 +89,14 @@ public abstract class FileUploadViewStrategy {
         .withId(FILEUPLOAD_SUBMIT_FORM_ID);
   }
 
-  private boolean hasUploadedFile(Params params) {
+  boolean hasUploadedFile(Params params) {
     return params.block().getQuestions().stream()
         .map(ApplicantQuestion::createFileUploadQuestion)
         .map(FileUploadQuestion::getFileKeyValue)
         .anyMatch(maybeValue -> maybeValue.isPresent());
   }
 
-  private boolean hasAtLeastOneRequiredQuestion(Params params) {
+  boolean hasAtLeastOneRequiredQuestion(Params params) {
     return params.block().getQuestions().stream().anyMatch(question -> !question.isOptional());
   }
 }
